Move Aserial into protocol files+ new algo for glider ratio
note : description for new glider ratio still not updated
