+++ openXsensor/oXs_out_multiplex.h
@@ -6,7 +6,9 @@
 #include "oXs_4525.h" // we need the airspeeddata struct
 #include "oXs_curr.h" // we need the currentdata struct
 #include "oXs_voltage.h" // we need the arduinodata struct
-#include "Aserial.h"
+//#include <Arduino.h>
+#include "oXs_general.h"
+
 
 #ifdef MULTIPLEX
 
@@ -114,12 +116,12 @@ struct t_mbAllData {
 #define SEND        3
 
 
-class OXS_OUT_FRSKY {
+class OXS_OUT {
   public:
 #ifdef DEBUG  
-    OXS_OUT_FRSKY(uint8_t pinTx,HardwareSerial &print);
+    OXS_OUT(uint8_t pinTx,HardwareSerial &print);
 #else
-    OXS_OUT_FRSKY(uint8_t pinTx);
+    OXS_OUT(uint8_t pinTx);
 #endif
     VARIODATA* varioData ;
     VARIODATA* varioData_2 ;
@@ -136,7 +138,6 @@ class OXS_OUT_FRSKY {
     HardwareSerial* printer;
 #endif
 
-// used by Multiplex protocol
     void formatAllMultiplexData() ; 
     uint8_t formatOneValue ( uint8_t currentFieldToSend ) ;
     void setMultiplexNewData(  uint16_t id, int32_t value  , uint8_t alarm) ;
@@ -162,6 +163,123 @@ extern uint8_t selectedVario ;
 extern volatile uint16_t RpmValue ;
 extern bool RpmAvailable ;
 #endif // MEASURE_RPM
+
+
+extern volatile uint8_t debug01 ;
+extern volatile uint8_t debug02 ;
+extern volatile uint8_t debug03 ;
+extern volatile uint8_t debug04 ;
+
+ 
+void setMultiplexNewData( struct t_sportData * volatile pdata, uint16_t id, uint32_t value ) ;
+void initMultiplexUart( struct t_mbAllData * volatile pdata ) ;
+
+
+extern volatile bool RpmSet ;
+extern volatile uint16_t RpmValue ;
+
+
+// UART's state.
+#define   IDLE               0        // Idle state, both transmit and receive possible.
+#define   TRANSMIT           1        // Transmitting byte.
+#define   TRANSMIT_STOP_BIT  2        // Transmitting stop bit.
+#define   RECEIVE            3        // Receiving byte.
+#define   TxPENDING          4
+#define   WAITING            5
+
+
+//This section chooses the correct timer values for Multiplex protocol = 38400 baud.
+// Assumes a 16MHz clock
+//#define TICKS2COUNT         (278*6)  // Ticks between two bits.
+//#define TICKS2WAITONE       (278*6)  // Wait one bit period.
+//#define TICKS2WAITONE_HALF  (416*6)  // Wait one and a half bit period.
+  #if F_CPU == 20000000L     // 20MHz clock                                                  
+    // Sinan: Not tested
+    #define TICKS2COUNTMULTIPLEX         (521)  // Ticks between two bits.
+    #define TICKS2WAITONEMULTIPLEX       (521)  // Wait one bit period.
+    #define TICKS2WAITONE_HALFMULTIPLEX  (781)    // Wait one and a half bit period.
+  #elif F_CPU == 16000000L   // 16MHz clock                                                  
+    #define TICKS2COUNTMULTIPLEX         (417)  // Ticks between two bits.
+    #define TICKS2WAITONEMULTIPLEX       (417)  // Wait one bit period.
+    #define TICKS2WAITONE_HALFMULTIPLEX  (625)    // Wait one and a half bit period.
+  #elif F_CPU == 8000000L    // 8MHz clock                                                   
+    #define TICKS2COUNTMULTIPLEX         (208)  // Ticks between two bits.
+    #define TICKS2WAITONEMULTIPLEX       (208)  // Wait one bit period.
+    #define TICKS2WAITONE_HALFMULTIPLEX  (313)    // Wait one and a half bit period.
+  #else
+    #error Unsupported clock speed
+  #endif
+
+
+
+//#define INTERRUPT_EXEC_CYCL   90       // Cycles to execute interrupt routines from interrupt.
+//#define INTERRUPT_EARLY_BIAS  32       // Cycles to allow of other interrupts.
+// INTERRUPT_EARLY_BIAS is to bias the sample point a bit early in case
+// the Timer 0 interrupt (5.5uS) delays the start bit detection
+  #if F_CPU == 20000000L     // 20MHz clock                                                  
+    #define INTERRUPT_EXEC_CYCL   112       // Cycles to execute interrupt routines from interrupt.
+    #define INTERRUPT_EARLY_BIAS  40       // Cycles to allow of other interrupts.
+  #elif F_CPU == 16000000L   // 16MHz clock                                                  
+    #define INTERRUPT_EXEC_CYCL   90       // Cycles to execute interrupt routines from interrupt.
+    #define INTERRUPT_EARLY_BIAS  32       // Cycles to allow of other interrupts.
+  #elif F_CPU == 8000000L    // 8MHz clock                                                   
+    #define INTERRUPT_EXEC_CYCL   45       // Cycles to execute interrupt routines from interrupt.
+    #define INTERRUPT_EARLY_BIAS  16       // Cycles to allow of other interrupts.
+  #else
+    #error Unsupported clock speed
+  #endif
+
+// this section define some delays used ; values can be used by any protocol
+  #if F_CPU == 20000000L     // 20MHz clock                                                  
+    #define DELAY_4000  ((uint16_t)4000.0 * 20.0 /16.0 )
+    #define DELAY_3500  ((uint16_t)3500.0 * 20.0 /16.0 )    
+    #define DELAY_2000  ((uint16_t)2000.0 * 20.0 /16.0 )
+    #define DELAY_1600  ((uint16_t)1600.0 * 20.0 /16.0 )    
+    #define DELAY_400  ((uint16_t)400.0 * 20.0 /16.0 )
+    #define DELAY_100  ((uint16_t)100.0 * 20.0 /16.0 )
+    
+  #elif F_CPU == 16000000L   // 16MHz clock                                                  
+    #define DELAY_4000 ((uint16_t) (1000L * 16) )     
+    #define DELAY_3500 ((uint16_t) (1000L * 16) )         
+    #define DELAY_2000 ((uint16_t) (1000L * 16) )     
+    #define DELAY_1600 ((uint16_t) (1000L * 16) )     
+    #define DELAY_400 ((uint16_t) (400 * 16) )     
+    #define DELAY_100 ((uint16_t) (100 * 16) )     
+  #elif F_CPU == 8000000L    // 8MHz clock                                                   
+    #define  DELAY_4000 ((uint16_t)4000L * 8 )
+    #define  DELAY_3500 ((uint16_t)3500L * 8 )    
+    #define  DELAY_2000 ((uint16_t)2000 * 8 )
+    #define  DELAY_1600 ((uint16_t)1600 * 8 )    
+    #define  DELAY_400 ((uint16_t)400 * 8 )
+    #define  DELAY_100 ((uint16_t)100 * 8 )    
+  #else
+    #error Unsupported clock speed
+  #endif
+
+#define TCCR             TCCR1A             //!< Timer/Counter Control Register
+#define TCCR_P           TCCR1B             //!< Timer/Counter Control (Prescaler) Register
+#define OCR              OCR1A              //!< Output Compare Register
+#define EXT_IFR          EIFR               //!< External Interrupt Flag Register
+#define EXT_ICR          EICRA              //!< External Interrupt Control Register
+
+#define TRXDDR  DDRD
+#define TRXPORT PORTD
+#define TRXPIN  PIND
+
+#define SET_TX_PIN( )    ( TRXPORT |= ( 1 << PIN_SERIALTX ) )
+#define CLEAR_TX_PIN( )  ( TRXPORT &= ~( 1 << PIN_SERIALTX ) )
+#define GET_RX_PIN( )    ( TRXPIN & ( 1 << PIN_SERIALTX ) )
+
+#define SET_TX_PIN_MB( ) ( TRXDDR &= ~( 1 << PIN_SERIALTX ) ) // in fact we put the port in high impedance because the rx has a pull up resistance
+#define CLEAR_TX_PIN_MB( ) ( TRXDDR |= ( 1 << PIN_SERIALTX ) ) // in fact we put the port in output mode. PORT is already set to 0 during initialisation
+
+#define MB_MAX_ADRESS  15
+
+// values used by the state ".active"
+#define NOT_ACTIVE     0
+#define NOT_AVAILABLE  1
+#define LOCKED         2
+#define AVAILABLE      3
 
 #endif // End of MULTIPLEX
 #endif // OXS_OUT_MULTIPLEX_h
