+++ openXsensor/version.oxs
@@ -1,1 +1,1 @@
-v6.0
+v6.1
